# youband-manager/module-albums/src/main/java/cz/muni/fi/pa165/modulealbums/rest/AlbumRestController.java
package cz.muni.fi.pa165.modulealbums.rest;

import cz.muni.fi.pa165.modulealbums.api.AlbumDto;
import cz.muni.fi.pa165.modulealbums.exceptions.ResourceNotFoundException;
import cz.muni.fi.pa165.modulealbums.facade.AlbumFacade;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping(path = "api/albums")
public class AlbumRestController {
    private final AlbumFacade albumFacade;

    @Autowired
    public AlbumRestController(AlbumFacade albumFacade) {
        this.albumFacade = albumFacade;
    }

    @GetMapping(path = "/{id}")
    public ResponseEntity<AlbumDto> findById(@PathVariable("id") Long id) {
        try {
            return ResponseEntity.ok(albumFacade.findById(id));
        } catch (ResourceNotFoundException ex) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }

    @GetMapping
    public ResponseEntity<List<AlbumDto>> getAll() {
        return ResponseEntity.ok(albumFacade.getAll());
    }

    @PostMapping
    public ResponseEntity<AlbumDto> createAlbum(@Valid @RequestBody AlbumDto albumDto){
        return ResponseEntity.ok(albumFacade.createAlbum(albumDto));
    }

    @PutMapping(path = "/{id}")
    public ResponseEntity<AlbumDto> updateAlbum(@PathVariable("id") Long id, @Valid @RequestBody AlbumDto albumDto) {
        try {
            return ResponseEntity.ok(albumFacade.updateAlbum(id, albumDto));
        } catch (ResourceNotFoundException ex) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }

    @DeleteMapping(path = "/{id}")
    public ResponseEntity<AlbumDto> deleteAlbum(@PathVariable("id") Long id) {
        try {
            albumFacade.deleteAlbum(id);
            return new ResponseEntity<>(HttpStatus.OK);
        } catch (ResourceNotFoundException ex) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }


}